Move some stuff to utils.rs
